CMT/requirements.txt
Django==1.9.5
django-crispy-forms==1.6.0
django-filter==0.13.0
django-guardian==1.4.4
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
httpie==0.9.3
Markdown==2.6.6
Pygments==2.1.3
PyJWT==1.4.0
requests==2.9.1
six==1.10.0
